[RENAME/#2] : d에서 Vote로 변경
