+++ app/src/main/java/com/sopt/sopkathon37android2/presentation/main/MainNavHost.kt
@@ -10,7 +10,7 @@ import androidx.navigation.navOptions
 import com.sopt.sopkathon37android2.presentation.a.navigation.aGraph
 import com.sopt.sopkathon37android2.presentation.b.navigation.bGraph
 import com.sopt.sopkathon37android2.presentation.c.navigation.cGraph
-import com.sopt.sopkathon37android2.presentation.d.navigation.dGraph
+import com.sopt.sopkathon37android2.presentation.vote.navigation.dGraph
 
 @Composable
 fun MainNavHost(
