+++ app/src/main/java/com/sopt/sopkathon37android2/presentation/vote/navigation/dNavigation.kt
@@ -1,4 +1,4 @@
-package com.sopt.sopkathon37android2.presentation.d.navigation
+package com.sopt.sopkathon37android2.presentation.vote.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavController
@@ -6,7 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.sopt.sopkathon37android2.core.navigation.Route
-import com.sopt.sopkathon37android2.presentation.d.DRoute
+import com.sopt.sopkathon37android2.presentation.vote.VoteRoute
 import kotlinx.serialization.Serializable
 
 fun NavController.navigateToD(navOptions: NavOptions? = null) {
@@ -15,7 +15,7 @@ fun NavController.navigateToD(navOptions: NavOptions? = null) {
 
 fun NavGraphBuilder.dGraph(paddingValues: PaddingValues) {
     composable<DScreen> {
-        DRoute(
+        VoteRoute(
             paddingValues = paddingValues,
         )
     }
